requirements.txt
numpy==1.22.3
Pillow==9.1.1
tk==0.1.0
